# docs/knowledge-base/api/openapi.json
{
  "components": {
    "schemas": {
...
      "DiagnosisCreate": {
        "properties": {
          "description": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "icd_code": {
            "maxLength": 16,
            "minLength": 1,
            "title": "Icd Code",
            "type": "string"
          }
        },
        "required": [
          "icd_code",
          "description"
        ],
        "title": "DiagnosisCreate",
        "type": "object"
      },
      "DiagnosisOut": {
        "properties": {
          "author_id": {
            "title": "Author Id",
            "type": "integer"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "encounter_id": {
            "title": "Encounter Id",
            "type": "integer"
          },
          "icd_code": {
            "title": "Icd Code",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "encounter_id",
          "author_id",
          "icd_code",
          "description"
        ],
        "title": "DiagnosisOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PrescribeRequest": {
        "description": "A doctor's request to prescribe within an encounter (story D1).",
        "properties": {
          "dose": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Dose",
            "type": "string"
          },
          "encounter_id": {
            "title": "Encounter Id",
            "type": "integer"
          },
          "medication_id": {
            "title": "Medication Id",
            "type": "integer"
          },
          "override_interaction": {
            "default": false,
            "description": "Acknowledge and proceed past a drug-interaction warning (\u00a75.4).",
            "title": "Override Interaction",
            "type": "boolean"
          },
          "refills": {
            "default": 0,
            "maximum": 12.0,
            "minimum": 0.0,
            "title": "Refills",
            "type": "integer"
          }
        },
        "required": [
          "encounter_id",
          "medication_id",
          "dose"
        ],
        "title": "PrescribeRequest",
        "type": "object"
      },
      "PrescriptionOut": {
        "properties": {
          "dose": {
            "title": "Dose",
            "type": "string"
          },
          "encounter_id": {
            "title": "Encounter Id",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "medication_id": {
            "title": "Medication Id",
            "type": "integer"
          },
          "patient_id": {
            "title": "Patient Id",
            "type": "integer"
          },
          "prescriber_id": {
            "title": "Prescriber Id",
            "type": "integer"
          },
          "refills": {
            "title": "Refills",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "encounter_id",
          "patient_id",
          "prescriber_id",
          "medication_id",
          "dose",
          "refills",
          "status"
        ],
        "title": "PrescriptionOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local-first medical portal. JSON API for accounts, appointments, clinical records, and prescriptions. The browser UI is server-rendered (Jinja2 + HTMX) by this same application.",
    "title": "HealthyVytals API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/encounters/{encounter_id}/diagnoses": {
      "post": {
        "operationId": "add_diagnosis_api_v1_encounters__encounter_id__diagnoses_post",
        "parameters": [
          {
            "in": "path",
            "name": "encounter_id",
            "required": true,
            "schema": {
              "title": "Encounter Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnosisCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosisOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add a diagnosis (doctor only)",
        "tags": [
          "clinical"
        ]
      }
    },
...
    "/api/v1/prescriptions": {
      "post": {
        "description": "Prescribe within an encounter after \u00a75.4 safety checks (story D1).\n\nA safety block raises ``UnsafePrescription`` \u2192 409 with a ``reason`` in the\nerror details (allergy | interaction | refill_cap). An interaction can be\nproceeded past with ``override_interaction=true``; an allergy cannot.",
        "operationId": "prescribe_api_v1_prescriptions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrescribeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prescribe a medication (doctor only, safety-checked)",
        "tags": [
          "prescriptions"
        ]
      }
    },
...
  }
}